refactor(rup): css class by semantic tag
