+++ src/app/modules/rup/components/elementos/deprecated/elementoDeRegistro.html
@@ -41,8 +41,7 @@
                     <div droppable [dropScope]="'vincular-registros-rup'" (onDrop)="vincularRegistros($event, registro)"
                          [dragOverClass]="'drag-target-border'" [ngClass]="{'relacionado': registro?.relacionadoCon}">
                         <div class="branch"></div>
-                        <div
-                             class="rup-card {{ (registro.esSolicitud) ? 'solicitud' : prestacionesService.getCssClass(registro.concepto) }}">
+                        <div class="rup-card {{ registro | semanticClass }}">
                             <div class="rup-header">
                                 <div class="icon-rup drag-handle" draggable
                                      [dragScope]="['orden-registros-rup', 'vincular-registros-rup', 'borrar-registros-rup']"
@@ -70,8 +69,7 @@
                                 </div>
                                 <div class="actions"
                                      *ngIf="!confirmarDesvincular[registro.id] && (!confirmarEliminar || (confirmarEliminar && indexEliminar != i) )">
-                                    <plex-badge
-                                                type="{{ (registro.esSolicitud) ? 'solicitud' : prestacionesService.getCssClass(registro.concepto) }}">
+                                    <plex-badge [type]="registro | semanticClass">
                                         {{!registro.esSolicitud ? registro.concepto.semanticTag : 'solicitud'}}
                                     </plex-badge>
                                     <plex-badge type="info ml-1"
