+++ src/app/modules/rup/components/elementos/rupers/seccionado/seccion.component.html
@@ -41,8 +41,7 @@
                     <div droppable [dropScope]="'vincular-registros-rup'" (onDrop)="vincularRegistros($event, registro)"
                          [dragOverClass]="'drag-target-border'" [ngClass]="{'relacionado': registro?.relacionadoCon}">
                         <div class="branch"></div>
-                        <div
-                             class="rup-card {{ (registro.esSolicitud) ? 'solicitud' : prestacionesService.getCssClass(registro.concepto) }}">
+                        <div class="rup-card {{ registro | semanticClass}}">
                             <div class="rup-header">
                                 <div class="icon-rup drag-handle" draggable
                                      [dragScope]="['orden-registros-rup', 'vincular-registros-rup', 'borrar-registros-rup']"
@@ -75,9 +74,8 @@
                                             Registro Privado
                                         </plex-badge>
                                     </ng-container>
-                                    <plex-badge
-                                                [type]="(registro.esSolicitud) ? 'solicitud' : prestacionesService.getCssClass(registro.concepto)">
-                                        {{registro.esSolicitud ?'solicitud':registro.concepto.semanticTag}}
+                                    <plex-badge [type]="registro | semanticClass">
+                                        {{ registro.esSolicitud ? 'solicitud': registro.concepto.semanticTag }}
                                     </plex-badge>
                                     <plex-badge type="info"
                                                 *ngIf='registro.esPrimeraVez && prestacion.estados[prestacion.estados.length - 1].tipo === "validada"'>
