+++ src/app/modules/rup/components/huds/vistaPrestacion.html
@@ -32,8 +32,7 @@
                 <h4>Registros de la consulta:</h4>
                 <ng-container *ngFor="let elemento of prestacion.ejecucion.registros">
 
-                    <div
-                         class="rup-card {{ elemento.esSolicitud ? 'solicitud' : servicioPrestacion.getCssClass(elemento.concepto) }}">
+                    <div class="rup-card {{ elemento | semanticClass }}">
                         <div class="rup-header">
                             <div class="icon-rup">
                                 <i class="adi {{ elemento | semanticIcon }}"></i>
