+++ src/app/modules/rup/services/prestaciones.service.ts
@@ -641,34 +641,6 @@ export class PrestacionesService {
     }
 
     /**
-     * Determina la clase a utilizar segun sematicTag de un concepto de SNOMED o si es turneable
-        *
-     * @param {any} conceptoSNOMED Concepto SNOMED a determinar tipo de icono
-     * @param {null} filtroActual Si estoy desde el buscador puedo indicar en que filtro estoy parado
-     * @returns string Clase a ser utilizado para estilizar las cards de RUP
-     * @memberof PrestacionesService
-     */
-    public getCssClass(conceptoSNOMED, esSolicitud) {
-        let clase = conceptoSNOMED.semanticTag;
-
-        if (conceptoSNOMED.esSolicitud || (typeof esSolicitud !== 'undefined' && esSolicitud)) {
-            clase = 'solicitud';
-        } else if (conceptoSNOMED.semanticTag === 'régimen/tratamiento') {
-            clase = 'regimen';
-        } else if (conceptoSNOMED.semanticTag === 'elemento de registro') {
-            clase = 'elementoderegistro';
-        } else if (conceptoSNOMED.semanticTag === 'evento') {
-            clase = 'hallazgo';
-        } else if (conceptoSNOMED.semanticTag === 'objeto físico' || conceptoSNOMED.semanticTag === 'medicamento clínico' || conceptoSNOMED.semanticTag === 'fármaco de uso clínico') {
-            clase = 'producto';
-        } else if (conceptoSNOMED.semanticTag === 'entidad observable') {
-            clase = 'procedimiento';
-        }
-
-        return clase;
-    }
-
-    /**
     * Devuelve el texto del informe del encuentro asociado al registro
     *
     * @param {any} paciente un paciente
@@ -747,39 +719,3 @@ export class PrestacionesService {
     }
 
 }
-
-
-export function getSemanticTag(concepto: ISnomedConcept, esSolicitud: boolean = false) {
-    if (esSolicitud) {
-        return 'solicitud';
-    } else {
-        switch (concepto.semanticTag) {
-            case 'hallazgo':
-            case 'evento':
-            case 'situación':
-                return 'hallazgo';
-
-            case 'trastorno':
-                return 'trastorno';
-
-            case 'procedimiento':
-            case 'entidad observable':
-            case 'régimen/tratamiento':
-                return 'procedimiento';
-
-            case 'producto':
-            case 'objeto físico':
-            case 'medicamento clínico':
-            case 'fármaco de uso clínico':
-                return 'producto';
-
-            case 'elemento de registro':
-                return 'elemento de registro';
-            default:
-                // No debería
-                return concepto.semanticTag;
-        }
-    }
-}
-
-
